requirements.txt
Django==1.11.12
pycrypto==2.6.1
pytz==2018.4
pyxdg==0.25
SecretStorage==2.3.1
six==1.10.0
psycopg2==2.7.4
djangorestframework==3.8.2
django-datatable==0.3.1
